+++ simple-stack-flow-masterdetail-fragments/src/main/java/com/example/stackmasterdetailfrag/SinglePaneFragmentStateChanger.java
@@ -30,7 +30,6 @@ public class SinglePaneFragmentStateChanger {
             fragmentTransaction.remove(noDetailsFragment);
         }
 
-
         for(Parcelable _oldPath : stateChange.getPreviousState()) {
             Paths.Path oldPath = (Paths.Path ) _oldPath;
             Fragment fragment = fragmentManager.findFragmentByTag(oldPath.getFragmentTag());